# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "乐动数据库服务API文档",
        "title": "Ledong DB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:31168",
    "basePath": "/api",
    "paths": {
...
        "/course/total": {
            "get": {
                "description": "根据开始时间查询课程列表，支持按会员编号过滤和分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "查询课程列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始时间，格式：2006-01-02 15:04:05",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "会员编号，可选",
                        "name": "number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，从1开始，默认1",
                        "name": "pageNum",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.CoursePage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/prepaidCard/course/create": {
            "post": {
                "description": "创建新课程，支持添加会员和消费记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "创建课程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "startTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "教练名称",
                        "name": "coachName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float32",
                        "description": "课程时长",
                        "name": "spendingTime",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "场地名称",
                        "name": "courtName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "描述",
                        "name": "descript",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "课程类型",
                        "name": "courseType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "会员JSON对象",
                        "name": "membersObj",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "是否成人课程",
                        "name": "isAdult",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/prepaidCard/course/duplicate": {
            "post": {
                "description": "检查课程列表中是否有重复的课程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "课程重复检查",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "description": "课程参数列表",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {}
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/prepaidCard/course/notify": {
            "post": {
                "description": "发送课程通知短信",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "课程通知",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "courseId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/prepaidCard/course/sm": {
            "post": {
                "description": "发送自定义短信",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "发送课程短信",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "短信参数",
                        "name": "params",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/prepaidCard/course/trial/{id}": {
            "post": {
                "description": "将体验课状态加1（从未成单变为成单）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "更新体验课状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/prepaidCard/course/{id}": {
            "delete": {
                "description": "删除课程并退还所有会员的消费",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "删除课程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/prepaidCard/course/{id}/{member}": {
            "delete": {
                "description": "从课程中移除指定会员，并退还消费",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "课程"
                ],
                "summary": "移除课程成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "会员编号",
                        "name": "member",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，0表示成功",
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.Coach": {
            "type": "object",
            "properties": {
                "charges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Charge"
                    }
                },
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Course"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "model.Course": {
            "type": "object",
            "properties": {
                "coach": {
                    "$ref": "#/definitions/model.Coach"
                },
                "coach_id": {
                    "type": "integer"
                },
                "course_type": {
                    "description": "-2体验课未成单,-1体验课成单,0订场，1班课，2私教",
                    "type": "integer"
                },
                "court": {
                    "description": "关联关系",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Court"
                        }
                    ]
                },
                "court_id": {
                    "description": "外键字段",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_adult": {
                    "description": "1=成人课程, 0=儿童课程",
                    "type": "integer"
                },
                "member": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PrepaidCard"
                    }
                },
                "notified": {
                    "type": "integer"
                },
                "spend": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Spend"
                    }
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "model.Court": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Course"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.PrepaidCard": {
            "type": "object",
            "properties": {
                "adults": {
                    "type": "integer"
                },
                "annual_count": {
                    "type": "number"
                },
                "annual_expire_time": {
                    "type": "string"
                },
                "charges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Charge"
                    }
                },
                "courses": {
                    "description": "关联关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Course"
                    }
                },
                "court": {
                    "type": "string"
                },
                "equivalent_balance": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "rest_charge": {
                    "type": "number"
                },
                "spends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Spend"
                    }
                },
                "times_count": {
                    "type": "number"
                },
                "times_expire_time": {
                    "type": "string"
                },
                "younths": {
                    "type": "integer"
                }
            }
        },
        "model.Spend": {
            "type": "object",
            "properties": {
                "annual_times": {
                    "type": "number"
                },
                "charge": {
                    "description": "消费金额",
                    "type": "number"
                },
                "description": {
                    "description": "注意：保持为float32类型",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "quantities": {
                    "type": "integer"
                },
                "times": {
                    "type": "number"
                }
            }
        },
...
        "service.CoursePage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Course"
                    }
                },
                "first": {
                    "type": "boolean"
                },
                "last": {
                    "type": "boolean"
                },
                "number": {
                    "type": "integer"
                },
                "numberOfElements": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
    }
}